Extract DataProfileScan to a separate module
